requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==2.0.0
passlib==1.7.4
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
tzdata==2023.3
Werkzeug==2.2.2
WTForms==3.0.1
